Move alerts to App.vue and fix closing causing multiple closes
